# gmc_api_gateway/docs/swagger.yaml
basePath: /gmcapi/v2
definitions:
...
  model.EVENT:
    properties:
      cluster:
        type: string
      eventTime:
        type: string
      kind:
        type: string
      message:
        type: string
      name:
        type: string
      namespace:
        type: string
      reason:
        type: string
      type:
        type: string
    type: object
  model.Error:
    properties:
      data:
        description: Email    string `json:"email"`
        type: string
      status:
        type: string
    type: object
...
  model.ROLE:
    properties:
      annotations: {}
      cluster:
        type: string
      createAt:
        type: string
      label: {}
      name:
        type: string
      project:
        type: string
      rules: {}
      user:
        type: string
      workspace:
        type: string
    type: object
  model.SECRET:
    properties:
      annotations: {}
      clusterName:
        type: string
      createAt:
        type: string
      data: {}
      dataCnt:
        type: integer
      events:
        items:
          $ref: '#/definitions/model.EVENT'
        type: array
      label: {}
      name:
        type: string
      namespace:
        type: string
      ownerReferences: {}
      type:
        type: string
      user:
        type: string
      workspaceName:
        type: string
    type: object
...
host: 101.79.1.138:8012
info:
  contact:
    name: GM-Center
    url: https://gedge-platform.github.io/gm-center/
  description: This is a Gedge GM-Center Swagger API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gedge GM-Center Swagger API
  version: "2.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: get Role List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ROLE'
      security:
      - Bearer: []
      summary: Show List Role
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Role
      parameters:
      - description: Role Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ROLE'
      security:
      - Bearer: []
      summary: Create Role
      tags:
      - Kubernetes
  /roles/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Role
      parameters:
      - description: name of the Role
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Role
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Role Details
      parameters:
      - description: name of the Role
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ROLE'
      security:
      - Bearer: []
      summary: Show detail Role
      tags:
      - Kubernetes
  /secrets:
    get:
      consumes:
      - application/json
      description: get Secret List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SECRET'
      security:
      - Bearer: []
      summary: Show List Secret
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Secret
      parameters:
      - description: Secret Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SECRET'
      security:
      - Bearer: []
      summary: Create Secret
      tags:
      - Kubernetes
  /secrets/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Secret
      parameters:
      - description: name of the Secret
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Secret
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Secret Details
      parameters:
      - description: name of the Secret
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SECRET'
      security:
      - Bearer: []
      summary: Show detail Secret
      tags:
      - Kubernetes
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '"Type \"Bearer \" and then your API Token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"